Move orders query to query builders
